Add some generation tests
